+++ src/sqlfluff/core/linter/linting_result.py
@@ -8,7 +8,6 @@ from typing_extensions import Literal
 
 from sqlfluff.core.errors import CheckTuple
 from sqlfluff.core.linter.linted_dir import LintedDir, LintingRecord
-from sqlfluff.core.linter.linted_file import TMP_PRS_ERROR_TYPES
 from sqlfluff.core.timing import RuleTimingSummary, TimingSummary
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -94,12 +93,6 @@ class LintingResult:
             buff += path.get_violations(**kwargs)
         return buff
 
-    def violation_dict(self, **kwargs) -> dict:
-        """Return a dict of paths and violations."""
-        return self.combine_dicts(
-            *(path.violation_dict(**kwargs) for path in self.paths)
-        )
-
     def stats(self, fail_code: int, success_code: int) -> Dict[str, Any]:
         """Return a stats dictionary of this result."""
         all_stats: Dict[str, Any] = dict(files=0, clean=0, unclean=0, violations=0)
@@ -147,50 +140,32 @@ class LintingResult:
         # Iterate through all the files to get rule timing information so
         # we know what headings we're going to need.
         rule_codes: Set[str] = set()
-        file_timing_dicts: Dict[str, dict] = {}
-        for dir in self.paths:
-            for file in dir.files:
-                if not file.timings:  # pragma: no cover
+        for path in self.paths:
+            for record in path.as_records():
+                if "timings" not in record:  # pragma: no cover
                     continue
-                file_timing_dicts[file.path] = file.timings.get_rule_timing_dict()
-                rule_codes.update(file_timing_dicts[file.path].keys())
+                rule_codes.update(record["timings"].keys())
 
         with open(filename, "w", newline="") as f:
             writer = csv.DictWriter(
-                f, fieldnames=meta_fields + timing_fields + sorted(rule_codes)
+                # Metadata first, then step timings and then _sorted_ rule codes.
+                f,
+                fieldnames=meta_fields + timing_fields + sorted(rule_codes),
             )
 
+            # Write the header
             writer.writeheader()
 
-            for dir in self.paths:
-                for file in dir.files:
-                    if not file.timings:  # pragma: no cover
+            for path in self.paths:
+                for record in path.as_records():
+                    if "timings" not in record:  # pragma: no cover
                         continue
+
                     writer.writerow(
                         {
-                            "path": file.path,
-                            "source_chars": (
-                                len(file.templated_file.source_str)
-                                if file.templated_file
-                                else ""
-                            ),
-                            "templated_chars": (
-                                len(file.templated_file.templated_str)
-                                if file.templated_file
-                                else ""
-                            ),
-                            "segments": (
-                                file.tree.count_segments(raw_only=False)
-                                if file.tree
-                                else ""
-                            ),
-                            "raw_segments": (
-                                file.tree.count_segments(raw_only=True)
-                                if file.tree
-                                else ""
-                            ),
-                            **file.timings.step_timings,
-                            **file_timing_dicts[file.path],
+                            "path": record["filepath"],
+                            **record["statistics"],  # character and segment lengths.
+                            **record["timings"],  # step and rule timings.
                         }
                     )
 
@@ -201,9 +176,11 @@ class LintingResult:
         This method is useful for serialization as all objects will be builtin python
         types (ints, strs).
         """
-        return [
-            record for linted_dir in self.paths for record in linted_dir.as_records()
-        ]
+        return sorted(
+            (record for linted_dir in self.paths for record in linted_dir.as_records()),
+            # Sort records by filename
+            key=lambda record: record["filepath"],
+        )
 
     def persist_changes(self, formatter, fixed_file_suffix: str = "") -> dict:
         """Run all the fixes for all the files and return a dict."""
@@ -228,27 +205,11 @@ class LintingResult:
 
     def count_tmp_prs_errors(self) -> Tuple[int, int]:
         """Count templating or parse errors before and after filtering."""
-        total_errors = self.num_violations(
-            types=TMP_PRS_ERROR_TYPES,
-            filter_ignore=False,
-            filter_warning=False,
-        )
-        num_filtered_errors = 0
-        for linted_dir in self.paths:
-            for linted_file in linted_dir.files:
-                num_filtered_errors += linted_file.num_violations(
-                    types=TMP_PRS_ERROR_TYPES
-                )
+        total_errors = sum(path.num_unfiltered_tmp_prs_errors for path in self.paths)
+        num_filtered_errors = sum(path.num_tmp_prs_errors for path in self.paths)
         return total_errors, num_filtered_errors
 
     def discard_fixes_for_lint_errors_in_files_with_tmp_or_prs_errors(self) -> None:
         """Discard lint fixes for files with templating or parse errors."""
-        total_errors = self.num_violations(
-            types=TMP_PRS_ERROR_TYPES,
-            filter_ignore=False,
-            filter_warning=False,
-        )
-        if total_errors:
-            for linted_dir in self.paths:
-                for linted_file in linted_dir.files:
-                    linted_file.discard_fixes_if_tmp_or_prs_errors()
+        for path in self.paths:
+            path.discard_fixes_for_lint_errors_in_files_with_tmp_or_prs_errors()