+++ test/cli/commands_test.py
@@ -33,8 +33,6 @@ from sqlfluff.cli.commands import (
     rules,
     version,
 )
-from sqlfluff.core.parser import CommentSegment
-from sqlfluff.core.rules import BaseRule, LintFix, LintResult
 from sqlfluff.utils.testing.cli import invoke_assert_code
 
 # tomllib is only in the stdlib from 3.11+
@@ -352,7 +350,7 @@ def test__cli__command_render_stdin():
         # Check basic parsing, with the code only option
         (parse, ["-n", "test/fixtures/cli/passing_b.sql", "-c"]),
         # Check basic parsing, with the yaml output
-        (parse, ["-n", "test/fixtures/cli/passing_b.sql", "-c", "-f", "yaml"]),
+        (parse, ["-n", "test/fixtures/cli/passing_b.sql", "-c", "--format", "yaml"]),
         (parse, ["-n", "test/fixtures/cli/passing_b.sql", "--format", "yaml"]),
         # Check parsing with no output (used mostly for testing)
         (parse, ["-n", "test/fixtures/cli/passing_b.sql", "--format", "none"]),
@@ -473,7 +471,6 @@ def test__cli__command_lint_parse(command):
                     "test/fixtures/cli/fail_many.sql",
                     "-vvvvvvv",
                 ],
-                "y",
             ),
             1,
         ),
@@ -488,7 +485,6 @@ def test__cli__command_lint_parse(command):
                     "_fix",
                     "test/fixtures/cli/fail_many.sql",
                 ],
-                "y",
             ),
             1,
         ),
@@ -501,7 +497,6 @@ def test__cli__command_lint_parse(command):
                     "_fix",
                     "test/fixtures/cli/fail_many.sql",
                 ],
-                "y",
             ),
             1,
         ),
@@ -741,7 +736,7 @@ def generic_roundtrip_test(
     source_file,
     rulestring,
     final_exit_code=0,
-    force=True,
+    check=False,
     fix_input=None,
     fix_exit_code=0,
     input_file_encoding="utf-8",
@@ -769,8 +764,8 @@ def generic_roundtrip_test(
         args=[lint, ["--dialect=ansi", "--rules", rulestring, filepath]],
     )
     # Fix the file (in force mode)
-    if force:
-        fix_args = ["--rules", rulestring, "-f", filepath]
+    if check:
+        fix_args = ["--rules", rulestring, "--check", filepath]
     else:
         fix_args = ["--rules", rulestring, filepath]
     fix_args.append("--dialect=ansi")
@@ -822,7 +817,7 @@ def test__cli__command__fix(rule, fname):
             FROM my_schema.my_table
             where processdate ! 3
             """,
-            ["--force", "--fixed-suffix", "FIXED", "--rules", "CP01"],
+            ["--fixed-suffix", "FIXED", "--rules", "CP01"],
             None,
             1,
         ),
@@ -835,7 +830,7 @@ def test__cli__command__fix(rule, fname):
             where processdate {{ condition }}
             """,
             # Test the short versions of the options.
-            ["--force", "-x", "FIXED", "-r", "CP01"],
+            ["-x", "FIXED", "-r", "CP01"],
             None,
             1,
         ),
@@ -849,7 +844,7 @@ def test__cli__command__fix(rule, fname):
             where processdate ! 3  -- noqa: PRS
             """,
             # Test the short versions of the options.
-            ["--force", "-x", "FIXED", "-r", "CP01"],
+            ["-x", "FIXED", "-r", "CP01"],
             None,
             1,
         ),
@@ -861,7 +856,7 @@ def test__cli__command__fix(rule, fname):
             FROM my_schema.my_table
             WHERE processdate ! 3
             """,
-            ["--force", "--fixed-suffix", "FIXED", "--rules", "CP01"],
+            ["--fixed-suffix", "FIXED", "--rules", "CP01"],
             None,
             1,
         ),
@@ -873,7 +868,7 @@ def test__cli__command__fix(rule, fname):
             FROM my_schema.my_table
             WHERE processdate ! 3  --noqa: PRS
             """,
-            ["--force", "--fixed-suffix", "FIXED", "--rules", "CP01"],
+            ["--fixed-suffix", "FIXED", "--rules", "CP01"],
             None,
             0,
         ),
@@ -887,7 +882,6 @@ def test__cli__command__fix(rule, fname):
             where processdate ! 3
             """,
             [
-                "--force",
                 "--fixed-suffix",
                 "FIXED",
                 "--rules",
@@ -921,7 +915,7 @@ def test__cli__command__fix(rule, fname):
                 FROM my_schema.my_table
                 where processdate != 3""",
             ],
-            ["--force", "--fixed-suffix", "FIXED", "--rules", "CP01"],
+            ["--fixed-suffix", "FIXED", "--rules", "CP01"],
             [
                 None,
                 """SELECT my_col
@@ -956,12 +950,8 @@ def test__cli__fix_error_handling_behavior(sql, fix_args, fixed, exit_code, tmpd
         with pytest.raises(SystemExit) as e:
             fix(
                 fix_args
-                + [
-                    "-f",
-                    # Use the short dialect option
-                    "-d",
-                    "ansi",
-                ]
+                # Use the short dialect option
+                + ["-d", "ansi"]
             )
         assert exit_code == e.value.code
     for idx, this_fixed in enumerate(fixed):
@@ -1001,7 +991,6 @@ where processdate ! 3
     options = [
         "--dialect",
         "ansi",
-        "-f",
         "--fixed-suffix=FIXED",
         sql_path,
     ]
@@ -1040,58 +1029,6 @@ WHERE processdate ! 3
             )
     else:
         assert not os.path.isfile(fixed_path)
-
-
-_old_eval = BaseRule._eval
-_fix_counter = 0
-
-
-def _mock_eval(rule, context):
-    # For test__cli__fix_loop_limit_behavior, we mock BaseRule.crawl(),
-    # replacing it with this function. This function generates an infinite
-    # sequence of fixes without ever repeating the same fix. This causes the
-    # linter to hit the loop limit, allowing us to test that behavior.
-    if context.segment.is_type("comment") and "Comment" in context.segment.raw:
-        global _fix_counter
-        _fix_counter += 1
-        fix = LintFix.replace(
-            context.segment, [CommentSegment(f"-- Comment {_fix_counter}")]
-        )
-        return LintResult(context.segment, fixes=[fix])
-    else:
-        return _old_eval(rule, context)
-
-
-@pytest.mark.parametrize(
-    "sql, exit_code",
-    [
-        ("-- Comment A\nSELECT 1 FROM foo", 1),
-        ("-- noqa: disable=all\n-- Comment A\nSELECT 1 FROM foo", 0),
-    ],
-)
-@patch("sqlfluff.rules.layout.LT01.Rule_LT01._eval", _mock_eval)
-def test__cli__fix_loop_limit_behavior(sql, exit_code, tmpdir):
-    """Tests how "fix" behaves when the loop limit is exceeded."""
-    fix_args = ["--force", "--fixed-suffix", "FIXED", "--rules", "LT01"]
-    tmp_path = pathlib.Path(str(tmpdir))
-    filepath = tmp_path / "testing.sql"
-    filepath.write_text(textwrap.dedent(sql))
-    with tmpdir.as_cwd():
-        with pytest.raises(SystemExit) as e:
-            fix(
-                fix_args
-                + [
-                    "-f",
-                    "--dialect=ansi",
-                ]
-            )
-        assert exit_code == e.value.code
-    # In both parametrized test cases, no output file should have been
-    # created.
-    # - Case #1: Hitting the loop limit is an error
-    # - Case #2: "noqa" suppressed all lint errors, thus no fixes applied
-    fixed_path = tmp_path / "testingFIXED.sql"
-    assert not fixed_path.is_file()
 
 
 @pytest.mark.parametrize(
@@ -1229,13 +1166,13 @@ def test__cli__command_fix_stdin_error_exit_code(
         ("LT01", "test/fixtures/linter/indentation_errors.sql", "n", 1, 1),
     ],
 )
-def test__cli__command__fix_no_force(rule, fname, prompt, exit_code, fix_exit_code):
+def test__cli__command__fix_check(rule, fname, prompt, exit_code, fix_exit_code):
     """Round trip test, using the prompts."""
     with open(fname) as test_file:
         generic_roundtrip_test(
             test_file,
             rule,
-            force=False,
+            check=True,
             final_exit_code=exit_code,
             fix_input=prompt,
             fix_exit_code=fix_exit_code,
@@ -1282,7 +1219,24 @@ def test__cli__command_parse_serialize_from_stdin(serialize, write_file, tmp_pat
 @pytest.mark.parametrize(
     "sql,rules,expected,exit_code",
     [
-        ("select * from tbl", "CP01", [], 0),  # empty list if no violations
+        (
+            "select * from tbl",
+            "CP01",
+            [
+                {
+                    "filepath": "stdin",
+                    "statistics": {
+                        "raw_segments": 12,
+                        "segments": 24,
+                        "source_chars": 17,
+                        "templated_chars": 17,
+                    },
+                    # Empty list because no violations.
+                    "violations": [],
+                }
+            ],
+            0,
+        ),
         (
             "SElect * from tbl",
             "CP01",
@@ -1339,6 +1293,14 @@ def test__cli__command_parse_serialize_from_stdin(serialize, write_file, tmp_pat
                             ],
                         },
                     ],
+                    "statistics": {
+                        "raw_segments": 12,
+                        "segments": 24,
+                        "source_chars": 17,
+                        "templated_chars": 17,
+                    },
+                    # NOTE: There will be a timings section too, but we're not
+                    # going to test that.
                 }
             ],
             1,
@@ -1379,6 +1341,14 @@ def test__cli__command_parse_serialize_from_stdin(serialize, write_file, tmp_pat
                             ],
                         },
                     ],
+                    "statistics": {
+                        "raw_segments": 6,
+                        "segments": 11,
+                        "source_chars": 12,
+                        "templated_chars": 8,
+                    },
+                    # NOTE: There will be a timings section too, but we're not
+                    # going to test that.
                 }
             ],
             1,
@@ -1407,9 +1377,19 @@ def test__cli__command_lint_serialize_from_stdin(
     )
 
     if serialize == "json":
-        assert json.loads(result.output) == expected
+        result = json.loads(result.output)
+        # Drop any timing section (because it's less determinate)
+        for record in result:
+            if "timings" in record:
+                del record["timings"]
+        assert result == expected
     elif serialize == "yaml":
-        assert yaml.safe_load(result.output) == expected
+        result = yaml.safe_load(result.output)
+        # Drop any timing section (because it's less determinate)
+        for record in result:
+            if "timings" in record:
+                del record["timings"]
+        assert result == expected
     elif serialize == "none":
         assert result.output == ""
     else:
@@ -1923,29 +1903,6 @@ class TestProgressBars:
         assert "\rparsing: 0it" not in raw_output
         assert "\r\rlint by rules:" not in raw_output
 
-    def test_cli_lint_disabled_progress_bar_deprecated_option(
-        self, mock_disable_progress_bar: MagicMock
-    ) -> None:
-        """Same as above but checks additionally if deprecation warning is printed."""
-        result = invoke_assert_code(
-            args=[
-                lint,
-                [
-                    "--disable_progress_bar",
-                    "test/fixtures/linter/passing.sql",
-                ],
-            ],
-        )
-        raw_output = repr(result.output)
-
-        assert "\rpath test/fixtures/linter/passing.sql:" not in raw_output
-        assert "\rparsing: 0it" not in raw_output
-        assert "\r\rlint by rules:" not in raw_output
-        assert (
-            "DeprecationWarning: The option '--disable_progress_bar' is deprecated, "
-            "use '--disable-progress-bar'"
-        ) in raw_output
-
     def test_cli_lint_enabled_progress_bar(
         self, mock_disable_progress_bar: MagicMock
     ) -> None:
@@ -2027,44 +1984,6 @@ class TestProgressBars:
         assert r"\rrule LT01:" in raw_output
         assert r"\rrule CV05:" in raw_output
 
-    def test_cli_fix_disabled_progress_bar(
-        self, mock_disable_progress_bar: MagicMock
-    ) -> None:
-        """When progress bar is disabled, nothing should be printed into output."""
-        result = invoke_assert_code(
-            args=[
-                fix,
-                [
-                    "--disable-progress-bar",
-                    "test/fixtures/linter/passing.sql",
-                ],
-            ],
-        )
-        raw_output = repr(result.output)
-
-        assert (
-            "DeprecationWarning: The option '--disable_progress_bar' is deprecated, "
-            "use '--disable-progress-bar'"
-        ) not in raw_output
-
-    def test_cli_fix_disabled_progress_bar_deprecated_option(
-        self, mock_disable_progress_bar: MagicMock
-    ) -> None:
-        """Same as above but checks additionally if deprecation warning is printed."""
-        invoke_assert_code(
-            args=[
-                fix,
-                [
-                    "--disable_progress_bar",
-                    "test/fixtures/linter/passing.sql",
-                ],
-            ],
-            assert_output_contains=(
-                "DeprecationWarning: The option '--disable_progress_bar' is "
-                "deprecated, use '--disable-progress-bar'"
-            ),
-        )
-
 
 multiple_expected_output = """==== finding fixable violations ====
 == [test/fixtures/linter/multiple_sql_errors.sql] FAIL
@@ -2087,6 +2006,7 @@ def test__cli__fix_multiple_errors_no_show_errors():
         args=[
             fix,
             [
+                "--check",  # Run in check mode to get the confirmation.
                 "--disable-progress-bar",
                 "test/fixtures/linter/multiple_sql_errors.sql",
             ],
@@ -2104,7 +2024,6 @@ def test__cli__fix_multiple_errors_quiet_force():
             [
                 "--disable-progress-bar",
                 "test/fixtures/linter/multiple_sql_errors.sql",
-                "--force",
                 "--quiet",
                 "-x",
                 "_fix",
@@ -2117,7 +2036,7 @@ def test__cli__fix_multiple_errors_quiet_force():
     )
 
 
-def test__cli__fix_multiple_errors_quiet_no_force():
+def test__cli__fix_multiple_errors_quiet_check():
     """Test the fix --quiet option without --force."""
     invoke_assert_code(
         ret_code=0,
@@ -2126,6 +2045,7 @@ def test__cli__fix_multiple_errors_quiet_no_force():
             [
                 "--disable-progress-bar",
                 "test/fixtures/linter/multiple_sql_errors.sql",
+                "--check",  # Run in check mode to get the confirmation.
                 "--quiet",
                 "-x",
                 "_fix",
@@ -2152,6 +2072,7 @@ def test__cli__fix_multiple_errors_show_errors():
                 "--disable-progress-bar",
                 "--show-lint-violations",
                 "test/fixtures/linter/multiple_sql_errors.sql",
+                "--check",  # Run in check mode to get the confirmation.
             ],
         ],
     )
@@ -2189,6 +2110,7 @@ def test__cli__multiple_files__fix_multiple_errors_show_errors():
             fix,
             [
                 "--disable-progress-bar",
+                "--check",  # Run in check mode to get the confirmation.
                 "--show-lint-violations",
                 sql_path,
                 indent_path,