+++ test/fixtures/dialects/vertica/escape.sql
@@ -1,0 +1,13 @@
+SELECT E'\'';
+
+SELECT E'''';
+
+SELECT E'''\'';
+
+SELECT E'\\\'''';
+
+SELECT E'
+
+\\
+''
+\\';