+++ test/fixtures/dialects/vertica/single_quote.sql
@@ -1,0 +1,16 @@
+SELECT '';
+
+SELECT '''';
+
+SELECT '
+
+';
+
+SELECT '''aaa''';
+
+SELECT '
+''
+';
+
+-- TODO: Need to think how to fix this case in the dialect
+-- SELECT '\';